requirements.txt
# --- Core Frameworks ---
chainlit>=2.8.3
strands-agents>=1.13.0

# --- Models & LLM Integrations ---
ollama>=0.6.0
openai>=2.5.0

# --- AWS & Cloud Tooling ---
awscli>=1.42.55
boto3>=1.40.55
botocore>=1.40.55
mcp>=1.18.0

# --- Web & API Layer ---
python-dotenv>=1.1.1
uvicorn>=0.38.0
aiohttp>=3.13.1
aiofiles>=23.1.0
fastapi>=0.116.0

# --- Document Parsing & Retrieval ---
docling>=2.57.0
docxtpl>=0.20.1
opensearch-py>=3.0.0
python-docx>=1.2.0
reportlab>=4.4.4
sentence-transformers>=5.1.1


# --- Utilities & Tooling ---
docker>=7.1.0
numpy>=2.3.4
openpyxl>=3.1.5
pandas>=2.3.3
requests>=2.32.5
rich>=14.2.0
uv>=0.9.4

# --- Optional Integrations ---
minio>=7.2.18
